refactor layers list
