+++ napari/components/_layers_list/view/__init__.py
@@ -1,1 +1,1 @@
-from .main import QtLayersPanel
+from .main import QtLayersList
