+++ napari/components/_viewer/view/buttons.py
@@ -7,14 +7,14 @@ from .... import resources
 
 
 class QtLayersButtons(QFrame):
-    def __init__(self, layers):
+    def __init__(self, viewer):
         super().__init__()
 
-        self.layers = layers
-        self.deleteButton = QtDeleteButton(self.layers)
-        self.newMarkersButton = QtNewMarkersButton(self.layers)
-        self.newShapesButton = QtNewShapesButton(self.layers)
-        self.newLabelsButton = QtNewLabelsButton(self.layers)
+        self.viewer = viewer
+        self.deleteButton = QtDeleteButton(self.viewer)
+        self.newMarkersButton = QtNewMarkersButton(self.viewer)
+        self.newShapesButton = QtNewShapesButton(self.viewer)
+        self.newLabelsButton = QtNewLabelsButton(self.viewer)
 
         layout = QHBoxLayout()
         layout.addStretch(0)
@@ -27,15 +27,15 @@ class QtLayersButtons(QFrame):
 
 
 class QtDeleteButton(QPushButton):
-    def __init__(self, layers):
+    def __init__(self, viewer):
         super().__init__()
 
-        self.layers = layers
+        self.viewer = viewer
         self.setFixedWidth(28)
         self.setFixedHeight(28)
         self.setToolTip('Delete layers')
         self.setAcceptDrops(True)
-        self.clicked.connect(lambda: self.layers.remove_selected())
+        self.clicked.connect(lambda: self.viewer.layers.remove_selected())
 
     def dragEnterEvent(self, event):
         event.accept()
@@ -53,33 +53,33 @@ class QtDeleteButton(QPushButton):
 
 
 class QtNewMarkersButton(QPushButton):
-    def __init__(self, layers):
+    def __init__(self, viewer):
         super().__init__()
 
-        self.layers = layers
+        self.viewer = viewer
         self.setFixedWidth(28)
         self.setFixedHeight(28)
         self.setToolTip('New markers layer')
-        self.clicked.connect(lambda: self.layers.viewer._new_markers())
+        self.clicked.connect(lambda: self.viewer._new_markers())
 
 
 class QtNewShapesButton(QPushButton):
-    def __init__(self, layers):
+    def __init__(self, viewer):
         super().__init__()
 
-        self.layers = layers
+        self.viewer = viewer
         self.setFixedWidth(28)
         self.setFixedHeight(28)
         self.setToolTip('New shapes layer')
-        self.clicked.connect(lambda: self.layers.viewer._new_shapes())
+        self.clicked.connect(lambda: self.viewer._new_shapes())
 
 
 class QtNewLabelsButton(QPushButton):
-    def __init__(self, layers):
+    def __init__(self, viewer):
         super().__init__()
 
-        self.layers = layers
+        self.viewer = viewer
         self.setFixedWidth(28)
         self.setFixedHeight(28)
         self.setToolTip('New labels layer')
-        self.clicked.connect(lambda: self.layers.viewer._new_labels())
+        self.clicked.connect(lambda: self.viewer._new_labels())
